xdr: Add LedgerEntry.Normalize() method (#3361)
Add `LedgerEntry.Normalize()` method. Remove `transformEntry` in Horizon.

It's possible that ledger entry fetched from History Archives and Horizon DB are
equal but are not binary the same due to ledger entry extensions. Previously
normalization was done in `transformEntry` function but in #3356 a new function
`NormalizeClaimableBalance` was added. To keep the code consistent, I updated it
to support all ledger entry types and removed `transformEntry`.